Split behaviour to smaller pieces
